audio: fix LP64 DSP-channel back-link truncation (unk8 u32->uintptr_t) + null-deref-before-check
TDSPChannel::getLogicalChannel returns (TChannel*)unk8, but unk8 was u32 and the channel ptr
was passed as (u32)(uintptr_t)this -> 64-bit host pointer TRUNCATED -> garbage ptr -> SIGSEGV
in updatecallDSPChannel once the DSP pipeline was actually driven. Widen the channel-ptr path
u32->uintptr_t (unk8, alloc/allocate/free sigs, drop (u32) casts). Also move
updatecallDSPChannel's 'mgr = channel->unk4' after the null check (null-deref-before-check).
uintptr_t==u32 on GC (faithful). Adds SB_DBG_AUDIO startSeq trace.

+++ src/JSystem/JAudio/JASystem/JASDSPChannel.cpp
@@ -42,7 +42,7 @@ void TDSPChannel::init(u8 param)
 	unkC  = DSPInterface::getDSPHandle(param);
 }
 
-BOOL TDSPChannel::allocate(u32 param)
+BOOL TDSPChannel::allocate(uintptr_t param)
 {
 	if (!isUnk1One())
 		return false;
@@ -118,7 +118,7 @@ void TDSPChannel::initAll()
 	first = 0;
 }
 
-TDSPChannel* TDSPChannel::alloc(u32 param1, u32 param2)
+TDSPChannel* TDSPChannel::alloc(u32 param1, uintptr_t param2)
 {
 	if (param1)
 		return nullptr;
@@ -136,7 +136,7 @@ TDSPChannel* TDSPChannel::alloc(u32 param1, u32 param2)
 	return nullptr;
 }
 
-int TDSPChannel::free(TDSPChannel* channel, u32 param)
+int TDSPChannel::free(TDSPChannel* channel, uintptr_t param)
 {
 	if (!channel)
 		return -1;
